heif: initialize image pointer
In `heif/main.cpp` the image pointer is not initialized, which could
cause a memory access violation. `free_buffer` doesn't set the image
pointer to `NULL` and calling `free_buffer` two times will cause
a memory access violation.
Also, set an explicit emsdk tag, because the library won't build with
the latest version of the SDK.